+++ test/balance/balance_service_test.dart
@@ -6,12 +6,12 @@ import 'package:test/test.dart';
 import '../credentials.dart';
 
 void main() {
+  final Credentials credentials = Credentials.from(Platform.environment);
+
   group('BalanceService', () {
-    Credentials credentials;
     BalanceService balanceService;
 
     setUp(() {
-      credentials = Credentials.from(Platform.environment);
       balanceService = ApiBalanceService(credentials.API_TEST_KEY);
     });
 
@@ -28,5 +28,5 @@ void main() {
         expect(balance.type, isA<String>());
       });
     });
-  });
+  }, skip: !credentials.arePresent);
 }