+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.5
 
 - Use environment variables for the MB access keys for running the tests
+- Skip tests if environment variables aren't present
 - Code cleanup
 
 ## 1.3.4